fix(desktop, host-internal): Bedrock region helper 改用 renderer-safe 子路径
将 bedrockApiBaseFromRegion 等纯函数拆至 bedrock-region 子路径，settings-view 不再从 host-internal 主入口 value import，避免 Vite 预构建 AWS SDK 与 extensions 依赖链。
