+++ packages/host-internal/package.json
@@ -27,6 +27,10 @@
       "types": "./dist/openai-models.d.ts",
       "default": "./dist/openai-models.js"
     },
+    "./bedrock-region": {
+      "types": "./dist/bedrock-region.d.ts",
+      "default": "./dist/bedrock-region.js"
+    },
     "./reasoning-effort": {
       "types": "./dist/reasoning-effort.d.ts",
       "default": "./dist/reasoning-effort.js"
