+++ packages/host-internal/src/index.ts
@@ -16,6 +16,7 @@ export * from './file-rewind.js';
 export * from './marketplace.js';
 export * from './google-gemini-endpoints.js';
 export * from './bedrock-models.js';
+export * from './bedrock-region.js';
 export * from './model-provider-presets.js';
 export * from './openai-models.js';
 export * from './reasoning-effort.js';
